refactor: 파일 구조 수정
- modules 폴더 삭제
- View, Model 상위로 이동
